022-burette/magnum_opus_iii/requirements.txt
appdirs
quart
wheel
quart-openapi
flask-marshmallow
flask-sqlalchemy
flask-injector
marshmallow-sqlalchemy
sqlalchemy-utils
ariadne
